+++ R/personal_finances_fed.R
@@ -4,7 +4,7 @@
 #'
 #' @note For the elections prior to 2000, some information may be incomplete.
 #'
-#' @param year Election year. For this function, only the years 2006, 2010, and 2014 are available.
+#' @param year Election year. For this function, only the years 2006, 2010, 2014 and 2018 are available.
 #' 
 #' @param uf Federation Unit acronym (\code{character vector}).
 #' 
@@ -50,43 +50,53 @@
 
 personal_finances_fed <- function(year, uf = "all",  br_archive = FALSE, ascii = FALSE, encoding = "latin1", export = FALSE){
   
-  
   # Input tests
   test_encoding(encoding)
   test_fed_year(year)
   uf <- test_uf(uf)
   br_archive <- test_br(br_archive)
-  
-  # Downloads the data
-  dados <- tempfile()
-  sprintf("http://agencia.tse.jus.br/estatistica/sead/odsele/bem_candidato/bem_candidato_%s.zip", year) %>%
-    download.file(dados)
-  unzip(dados, exdir = paste0("./", year))
-  unlink(dados)
-  
-  message("Processing the data...")
-  
-  # Cleans the data
-  setwd(as.character(year))
-  banco <- juntaDados(uf, encoding, br_archive)
-  setwd("..")
-  unlink(as.character(year), recursive = T)
-  
-  # Changes variables names
-  if(year < 2014){
-    names(banco) <- c("DATA_GERACAO", "HORA_GERACAO", "ANO_ELEICAO", "DESCRICAO_ELEICAO",
-                      "SIGLA_UF", "SQ_CANDIDATO", "CD_TIPO_BEM_CANDIDATO", "DS_TIPO_BEM_CANDIDATO",
-                      "DETALHE_BEM", "VALOR_BEM", "DATA_ULT_TOTALIZACAO", "HORA_ULT_TOTALIZACAO")
-  }
     
- 
+  if(year > 2002){
+    
+    # Downloads the data
+    dados <- tempfile()
+    sprintf("http://agencia.tse.jus.br/estatistica/sead/odsele/bem_candidato/bem_candidato_%s.zip", year) %>%
+      download.file(dados)
+    unzip(dados, exdir = paste0("./", year))
+    unlink(dados)
+    
+    message("Processing the data...")
+    
+    # Cleans the data
+    setwd(as.character(year))
+    banco <- juntaDados(uf, encoding, br_archive)
+    setwd("..")
+    unlink(as.character(year), recursive = T)
+    
+    # Changes variables names
+    if(year < 2014){
+      names(banco) <- c("DATA_GERACAO", "HORA_GERACAO", "ANO_ELEICAO", "DESCRICAO_ELEICAO",
+                        "SIGLA_UF", "SQ_CANDIDATO", "CD_TIPO_BEM_CANDIDATO", "DS_TIPO_BEM_CANDIDATO",
+                        "DETALHE_BEM", "VALOR_BEM", "DATA_ULT_TOTALIZACAO", "HORA_ULT_TOTALIZACAO")
+    } else{
+      names(banco) <- c("DATA_GERACAO", "HORA_GERACAO", "ANO_ELEICAO", "COD_TIPO_ELEICAO", "NOME_TIPO_ELEICAO",      
+                         "COD_ELEICAO", "DESCRICAO_ELEICAO", "DATA_ELEICAO", "SIGLA_UF", "SIGLA_UE",               
+                         "NOME_UE", "SQ_CANDIDATO", "NUMERO_ORDEM_CANDIDATO", "CD_TIPO_BEM_CANDIDATO",
+                         "DS_TIPO_BEM_CANDIDATO", "DETALHE_BEM", "VALOR_BEM", "DT_ULTIMA_ATUALIZACAO",
+                         "HH_ULTIMA_ATUALIZACAO")
+    }
+    
+    
+    # Change to ascii
+    if(ascii == T) banco <- to_ascii(banco, encoding)
+    
+    # Export
+    if(export) export_data(banco)
+    
+    message("Done.\n")
+    return(banco)
+  } else{
+      message("Not disponible. Please, check the documentation and try again.\n")
+    }
+  }
   
-  # Change to ascii
-  if(ascii == T) banco <- to_ascii(banco, encoding)
-  
-  # Export
-  if(export) export_data(banco)
-  
-  message("Done.\n")
-  return(banco)
-}
